crypto-trading-backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.121.3"}
psycopg = {extras = ["binary"], version = "^3.2.13"}
ccxt = "^4.5.20"
python-dotenv = "^1.2.1"
pydantic-settings = "^2.12.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
sqlalchemy = "^2.0.44"
alembic = "^1.17.2"
psycopg2-binary = "^2.9.11"
celery = {extras = ["redis"], version = "^5.5.3"}
redis = "^5.0.0"
cryptography = "^46.0.3"
slowapi = "^0.1.9"
bcrypt = "^5.0.0"
sentry-sdk = "^2.45.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
